+++ processor/src/main/java/arez/processor/TrackedDescriptor.java
@@ -32,6 +32,7 @@ final class TrackedDescriptor
   @Nonnull
   private final String _name;
   private boolean _mutation;
+  private boolean _highPriority;
   private boolean _reportParameters;
   @Nullable
   private ExecutableElement _trackedMethod;
@@ -82,6 +83,7 @@ final class TrackedDescriptor
   }
 
   void setTrackedMethod( final boolean mutation,
+                         final boolean highPriority,
                          final boolean reportParameters,
                          @Nonnull final ExecutableElement method,
                          @Nonnull final ExecutableType trackedMethodType )
@@ -96,6 +98,7 @@ final class TrackedDescriptor
     else
     {
       _mutation = mutation;
+      _highPriority = highPriority;
       _reportParameters = reportParameters;
       _trackedMethod = Objects.requireNonNull( method );
       _trackedMethodType = Objects.requireNonNull( trackedMethodType );
@@ -160,7 +163,13 @@ final class TrackedDescriptor
     parameters.add( _componentDescriptor.getComponentNameMethodName() );
     parameters.add( "." + getName() );
     sb.append( _mutation );
-    sb.append( ", () -> super.$N() )" );
+    sb.append( ", () -> super.$N()" );
+    if ( _highPriority )
+    {
+      sb.append( "," );
+      sb.append( _highPriority );
+    }
+    sb.append( " )" );
     parameters.add( _onDepsChangedMethod.getSimpleName().toString() );
 
     builder.addStatement( sb.toString(), parameters.toArray() );