+++ processor/src/test/resources/input/com/example/autorun/HighPriorityAutorunModel.java
@@ -1,0 +1,13 @@
+package com.example.autorun;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Autorun;
+
+@ArezComponent
+public abstract class HighPriorityAutorunModel
+{
+  @Autorun( highPriority = true )
+  protected void doStuff()
+  {
+  }
+}