+++ processor/src/test/resources/input/com/example/computed/HighPriorityComputedModel.java
@@ -1,0 +1,14 @@
+package com.example.computed;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Computed;
+
+@ArezComponent
+public abstract class HighPriorityComputedModel
+{
+  @Computed( highPriority = true )
+  public long getTime()
+  {
+    return 0;
+  }
+}